requirements.txt
requests==2.20.1
Flask==1.0.2
Flask_Bootstrap==3.3.7.1
SQLAlchemy==1.2.15
pytesseract==0.2.6
Pillow==5.4.1
beautifulsoup4==4.7.1
